+++ packages/website/components/NavBar/index.tsx
@@ -10,6 +10,10 @@ import { ThemeContext } from "../../utils/themeContext";
 import RssButton from "../RssButton";
 import Item from "./item";
 import { encodeQuerystring } from "../../utils/encode";
+import {
+  HEADER_ACTION_LABELS,
+  ICON_ACTION_BUTTON_CLASS,
+} from "./a11y";
 export default function (props: {
   logo: string;
   logoDark: string;
@@ -67,8 +71,12 @@ export default function (props: {
           style={{ height: 56 }}
         >
           <div className="mx-4 flex items-center">
-            <div
-              className="cursor-pointer block md:hidden"
+            <button
+              type="button"
+              className={`${ICON_ACTION_BUTTON_CLASS} cursor-pointer block md:hidden`}
+              aria-label={HEADER_ACTION_LABELS.menu}
+              aria-expanded={props.isOpen}
+              aria-controls="nav-mobile"
               onClick={() => {
                 if (!props.isOpen) {
                   // 要打开
@@ -86,6 +94,7 @@ export default function (props: {
                   width="24"
                   height="24"
                   className="dark:text-dark fill-gray-600"
+                  aria-hidden="true"
                 >
                   <path
                     d="M904 160H120c-4.4 0-8 3.6-8 8v64c0 4.4 3.6 8 8 8h784c4.4 0 8-3.6 8-8v-64c0-4.4-3.6-8-8-8zM904 784H120c-4.4 0-8 3.6-8 8v64c0 4.4 3.6 8 8 8h784c4.4 0 8-3.6 8-8v-64c0-4.4-3.6-8-8-8zM904 472H120c-4.4 0-8 3.6-8 8v64c0 4.4 3.6 8 8 8h784c4.4 0 8-3.6 8-8v-64c0-4.4-3.6-8-8-8z"
@@ -93,7 +102,7 @@ export default function (props: {
                   ></path>
                 </svg>
               </span>
-            </div>
+            </button>
             {props.headerLeftContent == "siteLogo" && (
               <div className="hidden md:block transform translate-x-2">
                 <img
@@ -129,13 +138,15 @@ export default function (props: {
               })}
             </ul>
             <div className="flex nav-action">
-              <div
+              <button
+                type="button"
                 onClick={() => {
                   setShowSearch(true);
                   document.body.style.overflow = "hidden";
                 }}
-                title="搜索"
-                className="flex group transform hover:scale-110 transition-all select-none cursor-pointer"
+                title={HEADER_ACTION_LABELS.search}
+                aria-label={HEADER_ACTION_LABELS.search}
+                className={`${ICON_ACTION_BUTTON_CLASS} flex group transform hover:scale-110 transition-all select-none cursor-pointer`}
               >
                 <div className="flex items-center mr-0 sm:mr-2 hover:cursor-pointer   transition-all dark:text-dark fill-gray-600">
                   <svg
@@ -145,6 +156,7 @@ export default function (props: {
                     p-id="2305"
                     width="20"
                     height="20"
+                    aria-hidden="true"
                   >
                     <path
                       d="M789.804097 737.772047 742.865042 784.699846 898.765741 940.600545 945.704796 893.672746Z"
@@ -159,7 +171,7 @@ export default function (props: {
                 <div className="flex items-center ">
                   <KeyCard type="search"></KeyCard>
                 </div>
-              </div>
+              </button>
               <ThemeButton defaultTheme={props.defaultTheme} />
               {props.showRSS == "true" && (
                 <RssButton showAdminButton={props.showAdminButton == "true"} />